tools/requirements.txt
pygit2==1.7.2
bidict==0.22.0
Pillow==9.0.1
numpy==2.0.2
# changelogs
PyYaml==6.0.1 #Also the Linter.
beautifulsoup4==4.9.3
# ezdb
mysql-connector-python==8.0.33
